Rename i-PI driver so it's clear what it drives
also update description and autoformat
